Added State Variable Filter, Weird Reverse Effect, and Waveform Visualiser
In this update I have updated the user interface to more closely match my design in Illustrator. This involved editing a LookAndFeel class which used rotarySlider. I have also began development on the waveform visualiser aspect of this application, however it doesn't quite behave the way I want it to.

Through following a tutorial by TheAudioProgrammer on Youtube I've been able to implement the StateVariableFilter class in Juce. However, the next step with this is to visualise the curve of the filter based on the state, freq, and res parameters.

Finally, I have created a reverse effect with the delay which sounds pretty cool, however it is unreliable and creates quite a few audio artefacts which need to be fixed.